+++ fuzzytable/patterns/__init__.py
@@ -1,3 +1,7 @@
-from fuzzytable.patterns.fieldpattern import FieldPattern
+from fuzzytable.patterns.fieldpattern import \
+    FieldPattern, \
+    minratio_getter,\
+    minratio_setter, \
+    mode_getter
+from fuzzytable.main.string_analysis import mode_setter, DefaultValue
 from fuzzytable.patterns.sheetpattern import SheetPattern
-# from fuzzytable.patterns.datapattern import