+++ tests/test_v010_fieldpatterns.py
@@ -1,10 +1,9 @@
 import pytest
-from fuzzytable import exceptions, FuzzyTable
-from fuzzytable.main.fuzzytable import FuzzyTable
+from fuzzytable import exceptions, FuzzyTable, FieldPattern
 
 
 # 010/1 #####
-def test_fieldpatternerror():
+def test_10_1_fieldpatternerror():
     with pytest.raises(exceptions.InvalidFieldError):
         FuzzyTable(
             name='does not matter',
@@ -24,7 +23,7 @@ def test_fieldpatternerror():
     [42],
 ])
 #  010/2  #####
-def test_seek_but_no_fields(get_test_path, header_row_seek, fields):
+def test_10_2_seek_but_no_fields(get_test_path, header_row_seek, fields):
 
     # GIVEN a table whose headers are NOT in row 1...
     path = get_test_path('csv')
@@ -36,3 +35,30 @@ def test_seek_but_no_fields(get_test_path, header_row_seek, fields):
             header_row_seek=header_row_seek,
             fields=fields,
         )
+
+
+@pytest.mark.parametrize('searchterms_excludename,expected_matchedheader', [
+    pytest.param(False, 'first_name', id='include fieldname'),
+    pytest.param(True, 'last_name', id='exclude fieldname'),
+])
+#  010/3  #####
+def test_10_3_searchterms_excludename(searchterms_excludename, expected_matchedheader, firstlastnames):
+
+    # GIVEN a table with headers 'first_name' and 'last_name'...
+    path = firstlastnames.path
+
+    # WHEN user seeks header row without supplying needed or correct field_names...
+    field = FieldPattern(
+        name='first_name',
+        alias='last_name',
+        searchterms_excludename=searchterms_excludename,
+    )
+
+    ft = FuzzyTable(
+        path=path,
+        fields=field,
+    )
+
+    actual_matchedheader = ft.fields[0].header
+
+    assert actual_matchedheader == expected_matchedheader